# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HZ Navigator API",
    "version": "1.0.0",
    "description": "# HZ Navigator API Documentation\n\nHZ Navigator is a comprehensive HUBZone certification management platform that helps businesses maintain compliance with SBA HUBZone program requirements.\n\n## Features\n\n- **Authentication**: JWT-based authentication with role-based access control\n- **Business Management**: Create and manage HUBZone certified businesses\n- **Compliance Monitoring**: Real-time compliance tracking and alerts\n- **Document Management**: Secure document upload and OCR processing\n- **Contract Tracking**: Agency contract management and goal tracking\n- **HUBZone Maps**: Geographic verification with vector tile support\n- **Analytics**: Comprehensive reporting and analytics dashboards\n\n## Authentication\n\nMost endpoints require authentication via JWT Bearer token:\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Rate Limiting\n\n- Authentication endpoints: 5 requests per minute\n- Password reset: 3 requests per hour\n- Standard API: 100 requests per minute\n- File uploads: 10 requests per minute\n- Admin endpoints: 30 requests per minute",
    "contact": {
      "name": "HZ Navigator Support",
      "email": "support@hz-navigator.gov",
      "url": "https://hz-navigator.gov/support"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://hz-navigator.gov/terms"
    }
  },
  "servers": [
    {
...
    },
    {
...
    },
    {
...
  "paths": {
...
    "/api/contracts": {
      "get": {
        "summary": "Get contracts with filters",
        "tags": ["Contracts"],
        "parameters": [
          { "$ref": "#/components/parameters/AgencyIdHeader" },
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "fiscalYear", "in": "query", "schema": { "type": "integer" } },
          { "name": "isHubzoneContractor", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "List of contracts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Contract" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contract",
        "tags": ["Contracts"],
        "parameters": [{ "$ref": "#/components/parameters/AgencyIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateContractRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Contract" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "Contract": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "contractNumber": { "type": "string" },
          "title": { "type": "string" },
          "contractorName": { "type": "string" },
          "contractorUei": { "type": "string" },
          "isHubzoneContractor": { "type": "boolean" },
          "awardDate": { "type": "string", "format": "date" },
          "awardValue": { "type": "number" },
          "status": { "type": "string" }
        }
      },
      "CreateContractRequest": {
        "type": "object",
        "required": ["contractNumber", "title", "contractorName", "contractorUei", "awardDate", "awardValue"],
        "properties": {
          "contractNumber": { "type": "string" },
          "title": { "type": "string" },
          "contractorName": { "type": "string" },
          "contractorUei": { "type": "string" },
          "isHubzoneContractor": { "type": "boolean" },
          "awardDate": { "type": "string", "format": "date" },
          "awardValue": { "type": "number" }
        }
      },
...
    },
...
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
      "AgencyIdHeader": {
        "name": "X-Agency-Id",
        "in": "header",
        "schema": { "type": "string" }
      }
    }
  },
  "security": [{ "BearerAuth": [] }]
}